share service worker api
